Moved YtDownloader and SettingsManager classes to core/
